Добавляет функции взаимодействия c удалённым сервером
