+++ js/util.js
@@ -1,3 +1,5 @@
+import {ALERT_SHOW_TIME} from './constants.js';
+
 function getRandomInteger (min, max) {
   const lower = Math.ceil(Math.min(Math.abs(min), Math.abs(max)));
   const upper = Math.floor(Math.max(Math.abs(min), Math.abs(max)));
@@ -33,8 +35,28 @@ const onEscapeKeydown = (evt, onClose) => {
   }
 };
 
+const showAlert = (message) => {
+  const alert = document.createElement('div');
+  alert.style.zIndex = '100';
+  alert.style.position = 'absolute';
+  alert.style.left = '0';
+  alert.style.top = '0';
+  alert.style.right = '0';
+  alert.style.padding = '10px 3px';
+  alert.style.fontSize = '30px';
+  alert.style.textAlign = 'center';
+  alert.style.backgroundColor = 'red';
+  alert.textContent = message;
+  document.body.append(alert);
+
+  setTimeout(() => {
+    alert.remove();
+  }, ALERT_SHOW_TIME);
+};
+
+
 const body = document.querySelector('body');
 
 const previewImage = document.querySelector('.img-upload__preview img');
 
-export {getRandomInteger, createRandomId, getRandomArrayElement, isEscapeKey, onEscapeKeydown, body, previewImage};
+export {getRandomInteger, createRandomId, getRandomArrayElement, isEscapeKey, onEscapeKeydown, showAlert, body, previewImage};
